# base: make diff() use views and broadcasting