--- a/NEWS.md
+++ b/NEWS.md
@@ -29,6 +29,7 @@ Standard library changes
   * Channels now convert inserted values (like containers) instead of requiring types to match ([#29092]).
   * `range` can accept the stop value as a positional argument, e.g. `range(1,10,step=2)` ([#28708]).
   * `edit` can now be called on a module to edit the file that defines it ([#29636]).
+  * `diff` now supports arrays of arbitrary dimensionality and can operate over any dimension ([#29827]).
 
 Compiler/Runtime improvements
 -----------------------------