fixes #27957
fixes #15414 (didn't see this at the time of coding)

This uses array views and broadcasting to implement diff() in arbitrary dimension.
Added benefit is a significant performance increase for arrays with more than about 10 elements (at least in my tests).

```
x = rand(1000)
y = rand(1000,1000)
z = rand(10)
@btime diff($x);
@btime diff($y, dims=2);
@btime diff($z);
```
before:
```
919.474 ns (3 allocations: 7.98 KiB)
1.612 ms (4 allocations: 7.62 MiB)
41.718 ns (3 allocations: 208 bytes)
```
after:
```
504.201 ns (3 allocations: 8.03 KiB)
864.279 μs (4 allocations: 7.62 MiB)
41.598 ns (3 allocations: 256 bytes)
```