@timholy 2018-10-29
Sorry for one more request, but I think the docstring needs updating too. Does this need a NEWS item?