@stev47 2018-10-29
updated the docstring